+++ ui/src/main/java/com/jin/ui/ingredient/content/IngredientBody.kt
@@ -1,4 +1,4 @@
-package com.jin.honey.feature.ingredient.ui.content
+package com.jin.ui.ingredient.content
 
 import androidx.compose.animation.AnimatedVisibility
 import androidx.compose.foundation.clickable
@@ -29,7 +29,7 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import com.jin.honey.R
+import com.jin.ui.R
 import com.jin.domain.food.model.Ingredient
 
 @Composable
